Frontend: Split of some of our js to vendor component related files
